fix(ci): resolve lint, race, and frontend max-lines errors
Backend:
- forbidigo: exclude internal/mthub/ (legitimate time.Now for server timestamps)
- gocognit: extract helpers in strategy_active_watch.go, mt4/orders.go, mt5/orders.go
- revive: exclude internal/mthub/ from argument-limit (OMS writer proto params)
- unparam: remove unused symbol param from contractSize in risk/rules.go
- race: add mutex to mockSubCliPerSymbol and demand_subscribe_test mocks
- fix mt4 stream type to grpc.ServerStreamingClient after refactor

Frontend:
- react-hooks/rules-of-hooks: move useUIStore before conditional return in AppSidebar
- max-lines: extract formatSchedule to scheduleFormat.ts (ScheduleTable)
- max-lines: extract columns to myStrategiesColumns.tsx (MyStrategiesTable)
- max-lines: extract formatTime/manualTriggerStart to utils (LiveSchedulesTab)

+++ backend/internal/connect/strategy/live_context.go
@@ -249,7 +249,7 @@ func (s *StrategyExecutionServer) backfillSymbolInfo(cfg LiveStrategyConfig, lct
 	lctx.Point = param.PointValue.String()
 	lctx.Digits = param.Digits
 	lctx.ContractSize = param.ContractSize.String()
-	lctx.StopsLevel = int32(param.StopLevel)
+	lctx.StopsLevel = param.StopLevel
 }
 
 // backfillTickSymbolInfo populates Point/Digits/ContractSize/StopsLevel on
@@ -265,5 +265,5 @@ func (s *StrategyExecutionServer) backfillTickSymbolInfo(cfg LiveStrategyConfig,
 	tctx.Point = param.PointValue.String()
 	tctx.Digits = param.Digits
 	tctx.ContractSize = param.ContractSize.String()
-	tctx.StopsLevel = int32(param.StopLevel)
+	tctx.StopsLevel = param.StopLevel
 }
